Introduce a CKEditorContext
This groups the functionalities around CKEditor into a single class.
