+++ tests/behat/behat.yml
@@ -17,9 +17,10 @@ default:
       contexts:
         - Drupal\DrupalExtension\Context\BatchContext
         - Drupal\DrupalExtension\Context\ConfigContext
+        - Drupal\social\Behat\CKEditorContext
+        - Drupal\social\Behat\EmailContext
         - Drupal\social\Behat\PostContext
         - Drupal\social\Behat\FeatureContext
-        - Drupal\social\Behat\EmailContext
         - Drupal\social\Behat\SocialDrupalContext
         - Drupal\social\Behat\SocialMessageContext
         - Drupal\social\Behat\SocialMinkContext
