+++ tests/behat/features/bootstrap/FeatureContext.php
@@ -63,60 +63,6 @@ class FeatureContext extends RawMinkContext implements Context
     }
 
     /**
-     * Get the wysiwyg instance variable to use in Javascript.
-     *
-     * @param string
-     *   The instanceId used by the WYSIWYG module to identify the instance.
-     *
-     * @throws Exception
-     *   Throws an exception if the editor does not exist.
-     *
-     * @return string
-     *   A Javascript expression representing the WYSIWYG instance.
-     */
-    protected function getWysiwygInstance($instanceId) {
-      $instance = "CKEDITOR.instances['$instanceId']";
-      if (!$this->getSession()->evaluateScript("return !!$instance")) {
-        throw new \Exception(sprintf('The editor "%s" was not found on the page %s', $instanceId, $this->getSession()->getCurrentUrl()));
-      }
-      return $instance;
-    }
-
-    /**
-     * @When /^I fill in the "([^"]*)" WYSIWYG editor with "([^"]*)"$/
-     */
-    public function iFillInTheWysiwygEditor($locator, $text) {
-      $field = $this->getSession()->getPage()->findField($locator);
-
-      if (null === $field) {
-        throw new ElementNotFoundException($this->getDriver(), 'form field', 'id|name|label|value|placeholder', $locator);
-      }
-
-      $id = $field->getAttribute('id');
-      $instance = $this->getWysiwygInstance($id);
-      $this->getSession()->executeScript("$instance.setData(\"$text\");");
-    }
-
-    /**
-     * @When /^I click on the embed icon in the WYSIWYG editor$/
-     */
-    public function clickEmbedIconInWysiwygEditor() {
-
-      $cssSelector = 'a.cke_button__social_embed';
-
-      $session = $this->getSession();
-      $element = $session->getPage()->find(
-        'xpath',
-        $session->getSelectorsHandler()->selectorToXpath('css', $cssSelector)
-      );
-      if (null === $element) {
-        throw new \InvalidArgumentException(sprintf('Could not evaluate CSS Selector: "%s"', $cssSelector));
-      }
-
-      $element->click();
-    }
-
-    /**
      * @Then /^The iframe in the body description should have the src "([^"]*)"$/
      */
     public function iFrameInBodyDescriptionShouldHaveTheSrc($src) {
@@ -166,26 +112,6 @@ class FeatureContext extends RawMinkContext implements Context
       if (strpos($iframe_source, $src) === FALSE) {
         throw new \InvalidArgumentException(sprintf('The iframe source does not contain the src: "%s" it is however: "%s"', $src, $iframe_source));
       }
-    }
-
-    /**
-     * @When /^I click on the image icon in the WYSIWYG editor$/
-     */
-    public function clickImageIconInWysiwygEditor() {
-
-      $cssSelector = 'a.cke_button__drupalimage';
-
-      $session = $this->getSession();
-      $element = $session->getPage()->find(
-        'xpath',
-        $session->getSelectorsHandler()->selectorToXpath('css', $cssSelector)
-      );
-      if (null === $element) {
-        throw new \InvalidArgumentException(sprintf('Could not evaluate CSS Selector: "%s"', $cssSelector));
-      }
-
-      $element->click();
-
     }
 
     /**
